Dialogs and Toasts

+++ app/favorites/favorites.component.ts
@@ -8,6 +8,9 @@ import { DrawerPage } from '../shared/drawer/drawer.page';
 
 import { View } from 'ui/core/view';
 
+import { confirm } from "ui/dialogs";
+import { Toasty } from 'nativescript-toasty';
+
 @Component({
     selector: 'app-favorites',
     moduleId: module.id,
@@ -35,9 +38,34 @@ export class FavoritesComponent extends DrawerPage implements OnInit {
     }
 
     deleteFavorite(id: number) {
-        this.favoriteservice.deleteFavorite(id)
-            .subscribe(favorites => this.favorites = new ObservableArray(favorites),
-                errmess => this.errMess = errmess);
+        console.log('delete', id);
+    
+        let options = {
+            title: "Confirm Delete",
+            message: 'Do you want to delete Dish '+ id,
+            okButtonText: "Yes",
+            cancelButtonText: "No",
+            neutralButtonText: "Cancel"
+        };
+    
+        confirm(options).then((result: boolean) => {
+            if(result) {
+    
+              this.favorites = null;
+    
+              this.favoriteservice.deleteFavorite(id)
+                  .subscribe(favorites => { 
+                    const toast = new Toasty("Deleted Dish "+ id, "short", "bottom");
+                    toast.show();
+                    this.favorites = new ObservableArray(favorites);
+                  },
+                  errmess => this.errMess = errmess);
+            }
+            else {
+              console.log('Delete cancelled');
+            }
+        });
+    
     }
 
     public onCellSwiping(args: ListViewEventData) {
